backend/IA/JapinhaAPI/requirements.txt
pydantic
fastapi
uvicorn
psycopg2-binary
numpy
pandas
scikit-learn
matplotlib
python-dotenv
mongoengine